requirements.txt
astroid==2.9.3
asttokens==2.0.5
attrs==19.3.0
Automat==0.8.0
backcall==0.2.0
beautifulsoup4==4.11.1
blinker==1.4
certifi==2019.11.28
chardet==3.0.4
click==8.0.1
cloud-init==21.4
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
dbus-python==1.2.16
decorator==5.1.1
distlib==0.3.4
distro==1.4.0
distro-info===0.23ubuntu1
entrypoints==0.3
executing==0.8.3
filelock==3.6.0
Flask==2.0.1
Flask-DebugToolbar==0.11.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.1
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==4.11.3
incremental==16.10.1
iniconfig==1.1.1
ipython==8.1.1
isort==5.10.1
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.0.1
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.7.1
lxml==4.8.0
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mccabe==0.6.1
more-itertools==4.2.0
netifaces==0.10.4
oauthlib==3.1.0
packaging==21.3
parso==0.8.3
pexpect==4.6.0
pickleshare==0.7.5
platformdirs==2.5.1
pluggy==1.0.0
prompt-toolkit==3.0.28
psycopg2-binary==2.9.1
pure-eval==0.2.2
py==1.11.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
Pygments==2.11.2
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
pylint==2.12.2
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==3.0.7
pyrsistent==0.15.5
pyserial==3.4
pytest==7.1.0
python-apt==2.0.0+ubuntu0.20.4.7
python-debian===0.1.36ubuntu1
PyYAML==5.3.1
requests==2.22.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
six==1.14.0
sos==4.3
soupsieve==2.3.2.post1
SQLAlchemy==1.4.23
ssh-import-id==5.10
stack-data==0.2.0
systemd-python==234
toml==0.10.2
tomli==2.0.1
traitlets==5.1.1
Twisted==18.9.0
typing_extensions==4.1.1
ubuntu-advantage-tools==27.6
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
virtualenv==20.13.3
wadllib==1.3.3
wcwidth==0.2.5
Werkzeug==2.0.1
wrapt==1.13.3
zipp==3.7.0
zope.interface==4.7.1
